+++ eva/RateSample.h
@@ -19,8 +19,8 @@ struct RateSample
 
     int64_t    deliveryRate = 0; // B/ms = kB/s
     int64_t    interval = -1; //us
-    uint32_t   delivered = 0;
-    uint32_t   priorDelivered = 0;
+    int64_t    delivered = 0;
+    int64_t    priorDelivered = 0;
     Timestamp  priorTime;
     int64_t    sendElapsed = -1; //us
     int64_t    ackElapsed = -1;  //us
